fix(clippy): clear pre-push gate lints in the self-hosting test drivers
The feat branch accumulated committed-but-never-gate-pushed clippy debt in the self-hosting
test drivers, surfaced by the pre-push hook (which runs clippy under DEFAULT features and
--all-targets, where earlier --all-features runs cfg'd these files out). Cleared with no
behavior change:

- tests/selfhost_body.rs, tests/selfhost_parse.rs, tests/selfhost_codegen.rs: file-level
  allows for the throwaway-adapter style lints (type_complexity, too_many_arguments, and, in
  parse, single_match / manual_range_patterns / identity_op for the documentary
  `temp * 1024 + offset` record literals), each with a comment stating why.
- tests/selfhost_codegen.rs: drop an unnecessary `mut` on the reconstruct driver's `set`
  closure (it captures nothing, so it is `Fn`).
- tests/wide_shared_segment.rs: remove a redundant `usize -> usize` cast.

Gate verified locally: cargo clippy --workspace --all-targets -D warnings, cargo fmt --check,
cargo doc (docs.rs feature set, CI flags), and the markdown-link check all clean.

+++ tests/wide_shared_segment.rs
@@ -94,7 +94,7 @@ fn set_and_get_shared_reach_a_high_offset_slot() {
         .slots
         .iter()
         .position(|s| s.name == "d.n")
-        .expect("n slot") as usize;
+        .expect("n slot");
     assert!(n_slot > u16::MAX as usize, "n slot {n_slot} exceeds 65535");
 
     let arena = Arena::with_capacity(DEFAULT_ARENA_CAPACITY);
